refactor(sdk): rename storage to storage_legacy
