+++ crates/sdk/src/lib.rs
@@ -14,7 +14,8 @@ pub mod panic;
 #[cfg(not(feature = "std"))]
 pub mod rwasm;
 pub mod shared;
-pub mod storage;
+#[deprecated(note = "Use `fluentbase_sdk::storage` instead", since = "0.4.5-dev")]
+pub mod storage_legacy;
 
 pub use allocator::*;
 pub use fluentbase_codec as codec;
